move validation tests from core to own test file
